Tighten NAPI boundary, drop standalone bin, fuse build jobs
- Replace standalone `[[bin]]` (which only supported `script`
  subcommands and errored on real commands) with `examples/script.rs`;
  Makefile now invokes `cargo run --example script`.
- Redact absolute paths (cwd / resolved config path) from NAPI error
  messages — callers already know their cwd; avoid leaking filesystem
  layout into off-host logs.
- Simplify `start_indexer` env_map construction into a single iterator
  chain.
- Add `packages/cli/clippy.toml` banning `napi::Error`/`Result`/… via
  `disallowed_types`. Per-file exception: `#![allow]` at the top of
  `src/napi.rs`. The type ban alone is sufficient — any
  `#[napi_derive::napi]` function has to return `napi::Result<T>`.
- Header EnvSafe.res as vendored from rescript-envsafe v5.0.0.
- Fuse `build-linux-package` + `build-envio-package` CI jobs; the addon
  build dominates wall time and the envio package build has no Rust
  dependency, so one job saves a checkout + toolchain setup.
