+++ .github/workflows/build_and_verify.yml
@@ -49,10 +49,13 @@ jobs:
 
   # ── Build ───────────────────────────────────────────────────────────
 
-  # Build envio CLI binary for linux-x64 and wrap it in a platform npm package
-  build-linux-package:
+  # Build envio NAPI addon + wrap it as envio-linux-x64, and build the base
+  # envio npm package (ReScript compilation). Fused so we pay for toolchain
+  # setup + checkout once — the addon build dominates total time, and the
+  # envio package build has no Rust dependency.
+  build-envio-package:
     runs-on: ubuntu-22.04
-    timeout-minutes: 9
+    timeout-minutes: 12
     needs: authorize
 
     steps:
@@ -60,44 +63,10 @@ jobs:
         with:
           ref: ${{ github.event.pull_request.head.sha || github.sha }}
 
-      - name: Install toolchain
+      - name: Install Rust toolchain
         uses: dtolnay/rust-toolchain@stable
 
       - uses: ./.github/actions/cargo-cache
-
-      - name: Build envio NAPI addon
-        env:
-          SVM_TARGET_PLATFORM: linux-amd64
-        run: cargo build --release --lib
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v6
-        with:
-          node-version: 24
-
-      - name: Build platform npm package
-        run: |
-          node packages/build-envio/build-platform-package.ts \
-            --version "0.0.1-dev" --platform "linux" --arch "x64" --out "envio-linux-x64"
-          mkdir -p envio-linux-x64
-          cp target/release/libenvio.so envio-linux-x64/envio.node
-
-      - name: Upload linux-x64 platform npm package
-        uses: actions/upload-artifact@v7
-        with:
-          name: npm-pkg-envio-linux-x64
-          path: envio-linux-x64
-
-  # Build envio base npm package (ReScript compilation, package.json rewrite)
-  build-envio-package:
-    runs-on: ubuntu-latest
-    timeout-minutes: 9
-    needs: authorize
-
-    steps:
-      - uses: actions/checkout@v6
-        with:
-          ref: ${{ github.event.pull_request.head.sha || github.sha }}
 
       - uses: pnpm/action-setup@v5
         with:
@@ -110,14 +79,32 @@ jobs:
           node-version: 24
           cache: "pnpm"
 
+      - name: Build envio NAPI addon
+        env:
+          SVM_TARGET_PLATFORM: linux-amd64
+        run: cargo build --release --lib
+
       - name: Install dependencies
         run: pnpm install
+
+      - name: Build linux-x64 platform npm package
+        run: |
+          node packages/build-envio/build-platform-package.ts \
+            --version "0.0.1-dev" --platform "linux" --arch "x64" --out "envio-linux-x64"
+          mkdir -p envio-linux-x64
+          cp target/release/libenvio.so envio-linux-x64/envio.node
 
       - name: Build envio package
         run: node packages/build-envio/src/build-artifact.ts --version 0.0.1-dev --out envio-dist
 
-      - name: Verify artifact
+      - name: Verify envio artifact
         run: node packages/build-envio/src/verify-artifact.ts envio-dist
+
+      - name: Upload linux-x64 platform npm package
+        uses: actions/upload-artifact@v7
+        with:
+          name: npm-pkg-envio-linux-x64
+          path: envio-linux-x64
 
       - name: Upload envio npm package artifact
         uses: actions/upload-artifact@v7
@@ -174,7 +161,7 @@ jobs:
   # Template tests - verify envio templates can be initialized and built
   template-tests:
     runs-on: ubuntu-latest
-    needs: [build-linux-package, build-envio-package]
+    needs: build-envio-package
     timeout-minutes: 9
 
     steps:
@@ -191,7 +178,7 @@ jobs:
   # Requires DB services (postgres + hasura)
   scenarios-test:
     runs-on: ubuntu-latest
-    needs: [build-linux-package, build-envio-package]
+    needs: build-envio-package
     timeout-minutes: 9
 
     services:
@@ -269,7 +256,7 @@ jobs:
   # E2E test - runs envio dev with a real indexer scenario
   e2e-test:
     runs-on: ubuntu-latest
-    needs: [build-linux-package, build-envio-package]
+    needs: build-envio-package
     timeout-minutes: 9
 
     services:
