+++ .github/workflows/build_and_verify_pr.yml
@@ -17,48 +17,19 @@ env:
   FORCE_JAVASCRIPT_ACTIONS_TO_NODE24: true
 
 jobs:
-  build-linux-package:
+  # Build envio NAPI addon + envio-linux-x64 platform pkg + base envio npm
+  # package in one job. See build_and_verify.yml for rationale.
+  build-envio-package:
     runs-on: ubuntu-22.04
-    timeout-minutes: 9
+    timeout-minutes: 12
 
     steps:
       - uses: actions/checkout@v6
 
-      - name: Install toolchain
+      - name: Install Rust toolchain
         uses: dtolnay/rust-toolchain@stable
 
       - uses: ./.github/actions/cargo-cache
-
-      - name: Build envio NAPI addon
-        env:
-          SVM_TARGET_PLATFORM: linux-amd64
-        run: cargo build --release --lib
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v6
-        with:
-          node-version: 24
-
-      - name: Build platform npm package
-        run: |
-          node packages/build-envio/build-platform-package.ts \
-            --version "0.0.1-dev" --platform "linux" --arch "x64" --out "envio-linux-x64"
-          mkdir -p envio-linux-x64
-          cp target/release/libenvio.so envio-linux-x64/envio.node
-
-      - name: Upload linux-x64 platform npm package
-        uses: actions/upload-artifact@v7
-        with:
-          name: npm-pkg-envio-linux-x64
-          path: envio-linux-x64
-
-  build-envio-package:
-    runs-on: ubuntu-latest
-    timeout-minutes: 9
-    needs: build-linux-package
-
-    steps:
-      - uses: actions/checkout@v6
 
       - uses: pnpm/action-setup@v5
         with:
@@ -71,14 +42,32 @@ jobs:
           node-version: 24
           cache: "pnpm"
 
+      - name: Build envio NAPI addon
+        env:
+          SVM_TARGET_PLATFORM: linux-amd64
+        run: cargo build --release --lib
+
       - name: Install dependencies
         run: pnpm install
+
+      - name: Build linux-x64 platform npm package
+        run: |
+          node packages/build-envio/build-platform-package.ts \
+            --version "0.0.1-dev" --platform "linux" --arch "x64" --out "envio-linux-x64"
+          mkdir -p envio-linux-x64
+          cp target/release/libenvio.so envio-linux-x64/envio.node
 
       - name: Build envio package
         run: node packages/build-envio/src/build-artifact.ts --version 0.0.1-dev --out envio-dist
 
-      - name: Verify artifact
+      - name: Verify envio artifact
         run: node packages/build-envio/src/verify-artifact.ts envio-dist
+
+      - name: Upload linux-x64 platform npm package
+        uses: actions/upload-artifact@v7
+        with:
+          name: npm-pkg-envio-linux-x64
+          path: envio-linux-x64
 
       - name: Upload envio npm package artifact
         uses: actions/upload-artifact@v7
