+++ packages/cli/clippy.toml
@@ -1,0 +1,27 @@
+# Keep the NAPI surface confined to src/napi.rs.
+#
+# Clippy's `disallowed-*` lints are *crate-wide*; per-file scoping is achieved
+# by putting `#![allow(clippy::disallowed_types)]` at the top of the one file
+# allowed to use these symbols (src/napi.rs). Anywhere else in the crate,
+# `use napi::Error` / `-> napi::Result<...>` / etc. fails
+# `cargo clippy -- -D warnings` (CI: .github/workflows/build_and_verify.yml).
+#
+# Wildcards are NOT supported, so we enumerate the types actually used. If
+# napi exposes new ones we want to index on, add them here. The type ban is
+# enough to block NAPI bindings — `#[napi_derive::napi]` expands into code
+# that returns `napi::Result<T>`, so it can't be applied without tripping
+# this list. (We don't ban the macro itself via `disallowed-macros` because
+# clippy doesn't honor `#[allow]` on attribute-macro invocations — it would
+# make the exception in napi.rs impossible.)
+#
+# Stronger alternative (compile-time, not lint-time): split the NAPI shim
+# into its own crate that depends on the CLI lib. Deferred until the surface
+# grows beyond three functions.
+
+disallowed-types = [
+  { path = "napi::Error", reason = "NAPI types must stay in src/napi.rs" },
+  { path = "napi::Result", reason = "NAPI types must stay in src/napi.rs" },
+  { path = "napi::Status", reason = "NAPI types must stay in src/napi.rs" },
+  { path = "napi::JsString", reason = "NAPI types must stay in src/napi.rs" },
+  { path = "napi::Env", reason = "NAPI types must stay in src/napi.rs" },
+]
