+++ packages/cli/src/commands.rs
@@ -241,11 +241,10 @@ pub mod start {
             .to_string_lossy()
             .into_owned();
 
-        let mut env_map = serde_json::Map::new();
-        env_map.insert("ENVIO_CONFIG".to_string(), config_path.into());
-        for (k, v) in extra_env {
-            env_map.insert(k.clone(), v.clone().into());
-        }
+        let env_map: serde_json::Map<String, serde_json::Value> =
+            std::iter::once(("ENVIO_CONFIG".to_string(), config_path.into()))
+                .chain(extra_env.iter().map(|(k, v)| (k.clone(), v.clone().into())))
+                .collect();
 
         Ok(Command::new(
             "start-indexer",
